Use int instead of uint_64
ROSE returns SgTypedefType for a uin64_t, which is maybe confusing for a first practical. Let's just use a regular int.

+++ complex-types/quiz-practicals/const_input.cxx
@@ -1,7 +1,7 @@
 #include <cstdint>
 #include <cmath>
 
-const uint64_t foo = 632356;
+const int foo = 632356;
 
 void bar() {
 	const float PI = 3.1415926;
